♻️ Keep system role scopes in roles.server
