+++ packages/scms-server/src/backend/index.ts
@@ -38,7 +38,6 @@ export * from './work-version-subject.server.js';
 export * from './work-version-affiliations.server.js';
 export * from './loaders/works/contains.server.js';
 export * from './works/versions/cloneDraftWorkVersion.server.js';
-export * from './systemRoleDefaults.js';
 
 export * from './etl/index.js';
 export * from './loaders/index.js';
